requirements.txt
flask==2.2.5
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
imbalanced-learn==0.14.0
plotly==5.15.0
dash==2.14.0
dash-bootstrap-components==1.4.1
joblib==1.3.2
pyyaml==6.0.1
xgboost==1.7.6
lightgbm==4.1.0